Implement voxel object prototype instancing
